[ADD] (image) 실제 api통해서 DB로 저장 (contentId 입력받았을떄)

+++ src/main/java/modak/camping/opendata/OpendataController.java
@@ -15,7 +15,7 @@ public class OpendataController {
     @PostMapping(value = "/base")
     public ResponseEntity postBase(@RequestParam(required = false, defaultValue = "3") int rangePageNo,
                                    @RequestParam(required = false, defaultValue = "1000") int numOfRows) throws Exception {
-        opendataService.getOpenDataJson("base",rangePageNo, numOfRows, -1);
+        opendataService.getOpenDataJson("base",rangePageNo, numOfRows, -1L);
         return  ResponseEntity.ok("ok");
     }
 
@@ -25,7 +25,7 @@ public class OpendataController {
     }
 
     @PostMapping(value = "/image")
-    public ResponseEntity postImage(@RequestParam(required = false, defaultValue = "-1") Integer contentId) throws Exception {
+    public ResponseEntity postImage(@RequestParam(required = false, defaultValue = "-1") Long contentId) throws Exception {
         opendataService.getOpenDataJson("image",-1,-1,contentId);
         return ResponseEntity.ok("ok");
     }
